Fix unnecessary free in o2i_SM2CiphertextValue
The caller will crash when using the return value from
o2i_SM2CiphertextValue. Because the returned pointer to SM2CiphertextValue has been free,
regardless of success or failure in this method.

When assign cv to ret, cv should be set NULL just like the behaviour in
SM2_do_encrypt.