Fix SKF and SDF engines build error
When build dynamic library on Windows, build errors occur on SKF and SDF
engines build. Because when the engines are compiled into a
self-contained dll, bind_engine and v_check are needed which has been
exported in def file.

Add missing methods here to resolve the build error.

Because the engines are dummy implemented, the methods added here are
also dummy.