Fix building issues introduced in "Update sm2 sign/enc with z256 implementation"
After 4fa09e1f54fd22765c31fd013b24302c31673a0e, building GmSSL with default
options (cmake without any -D) results in the following build error:

```
[ 62%] Linking C executable bin/gmssl
/usr/bin/ld: bin/libgmssl.so.3.1: undefined reference to `sm2_z256_modn_add'
/usr/bin/ld: bin/libgmssl.so.3.1: undefined reference to `sm2_z256_modn_rand'
/usr/bin/ld: bin/libgmssl.so.3.1: undefined reference to `sm2_z256_one'
...
```

This happens because sm2_z256_* family was introduced alonsgide the existing
sm2_* implementation. While it should be a optional feature with
-DENABLE_SM2_Z256 to turn it on/off, in reality sm2_z256_sign.c replaced
sm2_sign.c in the src list, so building GmSSL without Z256 feature always
results in failure now. (With Z256 enabled, the building issue is gone, but
that is only a workaround).

This fixes the issue by add sm2_{,z256_}sign to src list if ENABLE_SM2_Z256 is
disabled/enabled, and uses a if check to skip a test function for sm2 that is
only buildable when Z256 is enabled.

Signed-off-by: Guoxin Pu <pugokushin@gmail.com>